revamped login
